Some cleanup and config file support.
Initial support for configuration files (Not working yet).
Done some cleanup like moving the Directory::Path to a separate file
and simplifying the Node class interface.
Remove the offmedia tool, which won't be needed when we have
configuration files working.
Detect presence of /usr/include/db4.6 and db4.7 in the configure script.