Rework the path cache.
This patch should make the pathcache noticeably more efficient,
specially for deep directory hierarchies.

Modify Database::getregisters and Database::Register::getattrs so
that they fill a container passed as a parameter instead of
passing the result back as a return value, which involves an
additional copy.

Add a countattrs method to Database::Register, and use it instead
of getting the whole attribute list at several filesystem
operations.